style: fix clippy warnings in qbt_v2 module (M168)
Three cleanups to satisfy `cargo clippy -- -D warnings`:
- auth.rs doc list item overindentation (5 → 3 spaces)
- preferences.rs unnecessary `as f64` cast (f64 → f64)
- response.rs collapsible if-let (nested → && let)

+++ crates/irontide-api/src/routes/qbt_v2/preferences.rs
@@ -84,7 +84,7 @@ impl From<&Settings> for QbtPreferences {
         };
 
         let (max_ratio, max_ratio_enabled) = match s.seed_ratio_limit {
-            Some(r) => (r as f64, true),
+            Some(r) => (r, true),
             None => (-1.0, false),
         };
